WIP: Using meaningful names, updating in-code documentation, and removing obsolete comments, duplicated and commented code improves code readability. To do: Add prints to indicate the progress of the tests for 1D and 2D cases.
